moved room generators to own subpackage
